move dtos to separate file
