+++ src/server/handlers.rs
@@ -42,61 +42,8 @@ use crate::{
     AppState,
 };
 
-#[derive(Serialize, Debug)]
-pub struct Tag {
-    pub name: String,
-    pub id: String,
-    pub count: i64,
-}
+use super::dtos::{TagDto, PerformerDto};
 
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct Performer {
-    pub name: String,
-    pub id: String,
-    pub scene_count: i64,
-    pub image_url: Option<String>,
-    pub tags: Vec<String>,
-    pub rating: Option<i64>,
-    pub favorite: bool,
-}
-
-#[derive(Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub enum VideoDto {
-    Stash {
-        id: String,
-        title: String,
-        image_url: String,
-        performers: Vec<String>,
-        marker_count: usize,
-        tags: BTreeSet<String>,
-        interactive: bool,
-        studio: Option<String>,
-        rating: Option<i64>,
-    },
-    LocalFile {
-        id: String,
-        file_name: String,
-        interactive: bool,
-    },
-}
-
-impl From<StashScene> for VideoDto {
-    fn from(scene: StashScene) -> Self {
-        VideoDto::Stash {
-            id: scene.id,
-            title: scene.title,
-            image_url: "TODO".into(),
-            performers: scene.performers,
-            marker_count: scene.markers.len(),
-            tags: scene.tags.into_iter().collect(),
-            interactive: scene.interactive,
-            studio: scene.studio,
-            rating: scene.rating,
-        }
-    }
-}
 
 #[derive(Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
@@ -190,12 +137,12 @@ pub fn add_api_key(url: &str, api_key: &str) -> String {
 }
 
 #[axum::debug_handler]
-pub async fn fetch_tags() -> Result<Json<Vec<Tag>>, AppError> {
+pub async fn fetch_tags() -> Result<Json<Vec<TagDto>>, AppError> {
     let api = StashApi::load_config().await?;
     let tags = api.find_tags().await?;
     let mut tags: Vec<_> = tags
         .into_iter()
-        .map(|t| Tag {
+        .map(|t| TagDto {
             name: t.name,
             id: t.id,
             count: t.scene_marker_count.unwrap_or_default(),
@@ -210,13 +157,13 @@ pub async fn fetch_tags() -> Result<Json<Vec<Tag>>, AppError> {
 }
 
 #[axum::debug_handler]
-pub async fn fetch_performers() -> Result<Json<Vec<Performer>>, AppError> {
+pub async fn fetch_performers() -> Result<Json<Vec<PerformerDto>>, AppError> {
     let config = Config::get().await?;
     let api = StashApi::from_config(&config);
     let performers = api.find_performers().await?;
     let mut performers: Vec<_> = performers
         .into_iter()
-        .map(|p| Performer {
+        .map(|p| PerformerDto {
             id: p.id,
             scene_count: p.scene_count.unwrap_or_default(),
             name: p.name,
@@ -459,7 +406,7 @@ pub async fn get_video(
 pub async fn persist_marker(
     state: State<Arc<AppState>>,
     Json(marker): Json<CreateMarker>,
-) -> Result<Json<DbMarker>, AppError> {
+) -> Result<Json<Marker>, AppError> {
     tracing::info!("saving marker {marker:?} to the database");
     let marker = state.database.persist_marker(marker).await?;
