+++ src/server/mod.rs
@@ -1,2 +1,3 @@
 pub mod handlers;
 pub mod static_files;
+pub mod dtos;
